// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-similarity-linking-selection-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseSimilarityLinkingSelectionAyahDto } from '../models/phrase-similarity-linking-selection-ayah-dto';
import { PhraseSimilarityLinkingSelectionQueryDto } from '../models/phrase-similarity-linking-selection-query-dto';
export interface PhraseSimilarityLinkingSelectionResponse {
  activeBuildId: string;
  ayahs: Array<PhraseSimilarityLinkingSelectionAyahDto>;
  query: PhraseSimilarityLinkingSelectionQueryDto;
  selectedAyahCount: number;
}